Add files via upload
Write a program to check whether the number is positive, negative, or zero. Additionally, if the number is positive, check if it is even or odd. eg: num=6 positive,even

write a program to add two numbers if only both are even numbers eg:n1=4 n2=6 //o/p=10 n1=5 n2=4 // not valid

Write a program with a num prints the multiplication table for that number from 1 to 10 using a loop. eg:n=5 output:5 * 1 = 5 5 * 2 = 10